forward_service: rety a failed forwarder call
Up to one failed-to-forward sub-query will be executed locally (on a
super-coordinator). This local execution is meant as a fallback for
a single forward_request that could not be sent to its destined
coordinator (e.g. due gossiper not reacting fast enough). Local
execution was chosen as the safest one - it does not require sending
data to another coordinator.